Update BodyState in think loop, not in motion loop.
BodyState is updated immediately after capturing the camera image so that it's as close to the state of the body at the time of capture as possible.

This change reduces the load on the motion thread a bit, and means that throughout each think cycle, the body state doesn't change.