FSMCondition supports a 'condition factory'.
When transition conditions are stateful, it normally does not make sense to persist that state between occurences of the transition's origin state. For example, timings or counts should be reset to zero each time a state is entered.

Use a condition factory when the condition needs to be re-built each time a state is entered.