Config system now keeps a reference to the configuration.json Document.
This allows dynamic settings (ie. camera settings) to be loaded later in the cycle and have the configured value applied when they're added to the Config class.
Introduced a flag to know whether we're initialising the config system or not, as we need to be able to modify readonly settings during the configuration phase.
Reusable setting JSON value parsing logic via new pure virtual SettingBase::setValueFromJson(Value*).