Added code for finding half a convex hull.
I want to experiment using this over the field edge points, to see if it gives a better shape that doesn't exclude obstructions such as opponents.

This should be less susceptible to noise, but may have other problems such as allowing red parts of opponents to be considered as the ball, or allowing large green areas above the field edge to raise the edge line over a wide distance.

Will try it and see!