DataStreamer now runs on its own thread.
This reduces load on the main ThinkThread and became especially important lately where RGB JPEG image sizes grew considerably, making images very laggy in Round Table.

With this change, it is like watching video in the browser, and that work is no longer being done on the think thread. Note that there seems to be a memory leak in Round Table that requires investigation. It has always been there, but is more apparent now that we are able to load images at a *much* faster rate.