Encode debug image as PNG for some image types.
PNG is much crisper, and encodes faster. Now, in "None" mode, you can
view nearly 30fps in Round Table which really gives a good sense of
what the robot's vision system is producing.

It should be possible to do an even better job of encoding the PNG data
as we know the colour palette.