Introduce our own team/player state objects.
We want a version that we can control, and will map incoming/outgoing messages via mitecom in between that format and ours, breaking a dependency on mitecom and decoupling things.

In future we may have more that one publisher/listener so that we can communicate within our team on different ports with different protocols. This is a step in that direction.