Speed callback of LookAround option has longer-running effect.
Previously, the speed provided by the callback only applied in the one
cycle. With this change, the speed will gradually increase after being
dropped.

The rationale for this is that if an object is seen for one cycle, the
head should stay slow for a while longer in case it is seen again.
If no object is seen, the head will gradually approach full speed again.
If an object is seen, the head will linger with it in view, to the level
specified by the speed callback.

We may need to tweak some of the values provided by speed callbacks now,
as they may be too low in places, resulting in very slow movement.