Revert 'Remove broken goalpost width test'
This reverts commit 3a3c2b6d8c92ef654360ac8369837bfc464f9a89.

This check is not perfect, but since removing it I have started to see
tiny goals on the sides of the ball. The upper threshold can be
increased to 10 in order to avoid ignoring goals that appear angled in
the camera frame until such a time as we have a better technique for
this.